Reduce size of metrics in memory

+++ backend/libraries/chat_events/src/lib.rs
@@ -17,10 +17,10 @@ pub use crate::events_map::*;
 pub use crate::message_content_internal::*;
 pub use crate::stable_storage::key::*;
 
-fn incr(counter: &mut u64) {
+fn incr(counter: &mut u32) {
     *counter = counter.saturating_add(1);
 }
 
-fn decr(counter: &mut u64) {
+fn decr(counter: &mut u32) {
     *counter = counter.saturating_sub(1);
 }
